requirements.txt
openai==1.34.0
python-dotenv==1.0.1
python-pptx==0.6.23